requirements.txt
tweepy==4.14.0
openai==1.3.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.25.2
textblob==0.17.1
vaderSentiment==3.3.2
python-dateutil==2.8.2
pytz==2023.3
sqlite3
json5==0.9.14
colorama==0.4.6
tqdm==4.66.1
